[WIP]make resource links page to work with legal and housing
