+++ client/src/App.js
@@ -3,6 +3,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 import HomePage from './layout/homepage/HomePage';
 import SearchPage from './layout/search/SearchPage';
+import ResourceLinksPage from './layout/resource-links/ResourceLinksPage';
 
 class App extends Component {
   render() {
@@ -11,6 +12,8 @@ class App extends Component {
         <Switch>
           <Route exact path='/' component={HomePage}/>
           <Route path='/search-offered' component={SearchPage}/>
+          <Route path='/legal' render={() => (<ResourceLinksPage filter='legal-help' title='Legal Resources' />)} />
+          <Route path='/housing-resources' render={() => (<ResourceLinksPage filter='housing' title='Housing Resources' />)} />
         </Switch>
       </Router>
     );
